startinsert in non-prepended vim as well
I will alelviate my muscle memory woes by always starting git in insert
mode, whether the ticket number has been prepended or not.

The problem with this is for when you're editing non-empty commit
messages. You never expect that to start in insert mode. That's just
weird. So (forthcoming functionality), when something needs to be
prepended, vim needs to be launched a certain way. When nothing is to be
preneded, there are still 2 different ways to launch vim: 1 for new,
empty commits, 1 for pre-existing, populated commits.